+++ lib/Component/Payum/Stripe/Api.php
@@ -17,8 +17,12 @@ use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetPaymentMethods
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetSubscriptions;
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\CancelSubscription;
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetConnectedAccounts;
+
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetWebhookEndpoints;
+use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\RetrieveWebhookEndpoint;
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\CreateWebhookEndpoint;
+use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\UpdateWebhookEndpoint;
+use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\DeleteWebhookEndpoint;
 
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetCoupons;
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\CreateCoupon;
@@ -33,16 +37,16 @@ final class Api
     const PRICE_ATTRIBUTE_KEY           = 'stripe_price_id';
     
     const STRIPE_EVENTS                 = [
-        'charge.succeeded',
-        'charge.failed',
-        'invoice.finalized',
-        'invoice.finalization_failed',
-        'invoice.paid',
-        'invoice.payment_failed',
-        'invoice.payment_succeeded',
-        'subscription_schedule.canceled',
-        'subscription_schedule.created',
-        'subscription_schedule.completed',
+        1 => 'charge.succeeded',
+        2 => 'charge.failed',
+        3 => 'invoice.finalized',
+        4 => 'invoice.finalization_failed',
+        5 => 'invoice.paid',
+        6 => 'invoice.payment_failed',
+        7 => 'invoice.payment_succeeded',
+        8 => 'subscription_schedule.canceled',
+        9 => 'subscription_schedule.created',
+        10 => 'subscription_schedule.completed',
     ];
     
     /** @var Gateway */
@@ -206,13 +210,49 @@ final class Api
         return isset( $availableWebhookEndpoints["data"] ) ? $availableWebhookEndpoints["data"] : [];
     }
     
+    public function retrieveWebhookEndpoint( string $id )
+    {
+        $webhookEndpoint = new \ArrayObject( ['id' => $id] );
+        $this->gateway->execute( $retrieveWebhookEndpointRequest = new RetrieveWebhookEndpoint( $webhookEndpoint ) );
+        
+        return $retrieveWebhookEndpointRequest->getFirstModel()->getArrayCopy();
+    }
+    
     public function createWebhookEndpoint( array $formData )
     {
+        $enabledEvents  = [];
+        foreach ( $formData['enabled_events'] as $event ) {
+            $enabledEvents[]    = self::STRIPE_EVENTS[$event];
+        }
+        
         $webhookEndpoint    = new \ArrayObject([
-            'enabled_events'    => $formData['enabled_events'],
+            'enabled_events'    => $enabledEvents,
             'url'               => $formData['url'],
         ]);
         $this->gateway->execute( new CreateWebhookEndpoint( $webhookEndpoint ) );
+    }
+    
+    public function updateWebhookEndpoint( array $formData )
+    {
+        $enabledEvents  = [];
+        foreach ( $formData['enabled_events'] as $event ) {
+            $enabledEvents[]    = self::STRIPE_EVENTS[$event];
+        }
+        
+        $webhookEndpoint    = new \ArrayObject([
+            'id'                => $formData['id'],
+            'enabled_events'    => $enabledEvents,
+            'url'               => $formData['url'],
+        ]);
+        $this->gateway->execute( new UpdateWebhookEndpoint( $webhookEndpoint ) );
+    }
+    
+    public function deleteWebhookEndpoint( $id )
+    {
+        $webhookEndpoint   = new \ArrayObject([
+            "id"    => $id,
+        ]);
+        $this->gateway->execute( new DeleteWebhookEndpoint( $webhookEndpoint ) );
     }
     
     public function getEvent( $eventId )