+++ lib/Form/Stripe/WebhookEndpointForm.php
@@ -4,6 +4,7 @@ use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
+use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 
@@ -14,20 +15,23 @@ class WebhookEndpointForm extends AbstractType
     public function buildForm( FormBuilderInterface $builder, array $options ): void
     {
         $builder
+            ->add( 'id', HiddenType::class, ['data' => $options['endpointId']] )
+            
+            ->add( 'selectedEvents', HiddenType::class, ['data' => \json_encode( $options['endpointEvents'] )] )
+        
             ->add( 'enabled_events', ChoiceType::class, [
                 'label'                 => 'vs_payment.template.payum_stripe_objects.enabled_events',
                 'translation_domain'    => 'VSPaymentBundle',
-                'choices'               => \array_combine( StripeApi::STRIPE_EVENTS, StripeApi::STRIPE_EVENTS ),
+                'choices'               => \array_flip( StripeApi::STRIPE_EVENTS ),
                 'multiple'              => true,
                 'required'              => false,
             ])
         
             ->add( 'url', TextType::class, [
-                'label' => 'vs_payment.template.payum_stripe_objects.url',
-                'translation_domain' => 'VSPaymentBundle',
-                'attr'  => [
-                    'placeholder' => 'vs_payment.template.payum_stripe_objects.url'
-                ],
+                'label'                 => 'vs_payment.template.payum_stripe_objects.url',
+                'attr'                  => ['placeholder' => 'vs_payment.template.payum_stripe_objects.url'],
+                'translation_domain'    => 'VSPaymentBundle',
+                'data'                  => $options['endpointUrl']
             ])
         ;
     }
@@ -39,6 +43,10 @@ class WebhookEndpointForm extends AbstractType
         $resolver
             ->setDefaults([
                 'csrf_protection'   => false,
+                
+                'endpointId'        => null,
+                'endpointEvents'    => [],
+                'endpointUrl'       => '',
             ])
         ;
     }